build: Rework test script to use go.mod files.
This reworks the logic in the test run script for determining which
modules to test to use the existence of a go.mod file instead of
attempting to use the result of go list.

The existing logic is to determine the root module via go list and then
test everything that has that as a prefix.  However, that is error prone
because indirect dependencies may depend on older versions of modules
that are no longer in the repo, but naturally still have the prefix that
is specified to test.

Searching for the go.mod files in the repo isn't prone to the
aforementioned issue and is generally more robust because it will always
be recent as of the current state of the repo and it still has the
benefit of dynamically finding all modules in the repo.