mixclient: Do not submit PRs holding client mutex
The client mutex was being held during the initial publishing of pair request
messages to prevent a situation where a submit errored but the peer was still
associated with the client, or the message was submitted to mixpool and sent
to other peers but our local peer had not yet been associated to the client at
the time an epoch tick occurred.  This should not be a situation we will
encounter anymore, since the PR submissions are spaced out intentionally to
avoid sending them close to the epoch.