[release-v2.0] mixclient: Use newest (fewest-PR) KEs to form alt sessions
The alternate session forming only incrementally removes PRs from the
currently considered PR set.  Even if a PR by a responsive peer is known, if
it was removed due to not passing the majority checks earlier, it will never
be used by our peers during this epoch.  With that in mind, we should only use
the most recent KEs with the lowest referenced PR counts when trying to form
an alternate session, as the additional PRs from earlier KEs will never be
reconsidered by that KE's identity.