addrmgr: Track network address types.
This commit adds support for determining and tracking the type of a
network address when constructed. While here, several functions are
refactored for easier reading and interpretability.

The types are first defined in network.go, and then constructed in
netaddress.go. The rest of the changes propagate outward from there.
Minimal changes were made to anything related to TorV2, because it will
be removed in a future commit.