blake256: Full redesign.
The current blake256 implementation has served well enough over the
years, however, the underlying implementation it was based on has a few
less than desirable qualities.  For example, it was not designed in a
way that is conducive to avoiding memory allocations nor exposing new
functionality because it only returns a standard library interface.  In
addition, the core compression code is not very well optimized,
particularly for smaller input sizes which are extremely common.

Consequently, this starts a series of commits which rewrite the package
from scratch to address the shortcomings while significantly optimizing
it along the way to include adding asm optimizations that takes
advantage of SIMD vector extensions (SSE2, SSE4.1, and AVX) on amd64
processors that supports them.  It does not require any dependencies
beyond the standard library.

Note that in order to help ease the review process, this commit only
implements the pure Go code along with some basic infrastructure to
support implementing the aforementioned asm optimizations while the
optimizations themselves are implemented in future commits.

The design philosophy of the new implementation places a strong emphasis
correctness, readability, and efficiency while also aiming to provide an
ergonomic API.

In order to avoid needing a new major module version, care has been
taken to expose all existing public functions, methods, and constants in
addition to the new concrete types and methods that expose the new
capabilities.  Naturally, callers using the old API will still be
subject to its limitations.  However, they will still gain the speed
benefits of the new better optimized implementations which are used
under the hood.

The following is an overview of some of the key highlights of the new
implementation:

- Enables zero allocations for rolling hashes
- Concrete rolling hashers (for BLAKE-224 and BLAKE-256) with
  convenience methods for directly writing unsigned integers of various
  sizes encoded as big or little endian as well as single bytes
- Ability to save and restore the intermediate state (also often called
  the midstate):
  - Same process state saving with a simple instance copy
  - Cross process state saving via serialization
- Provides asm implementations accelerated by SSE2, SSE4.1, and AVX that
  are dynamically selected based on processor support (future commits)
  - Supports the `purego` build tag to disable all asm code if desired
  - Supports environment variables to selectively disable extensions
- Hashing with salt
- Comprehensive test coverage with 100% branch coverage (future commits)
- Full suite of benchmarks for analyzing performance (future commits)
- Supporting documentation (future commits)
- Example usage for various scenarios (future commit)
- No dependencies other than the standard library