blake256: Add cpu feature detection.
This updates the internal compress package to detect relevant supported
CPU features for upcoming specialized implementations and arranges for
the exported Blocks function to automatically use the fastest available
supported hardware extensions that are not disabled.

It also updates the main package and compress package README.md files to
reflect the new capability.