blake256: Add specialized SSE2 implementation.
This adds a specialized assembly implementation that takes advantage of
SSE2 SIMD vector extensions on the amd64 architecture.

In order to significantly ease the review and implementation of the
assembly code that takes advantage of the amd64 SIMD vector extensions,
it is generated with Go via avo and includes extensive comments.  This
means only the Go code that generates the asm needs to be asserted for
correctness versus needing to review thousands of lines of pure asm
code.  It also means most of the complexity is handled in Go at compile
time and Go control structures can be used without the associated
overhead in the generated asm since that is completely unrolled.

Also, since the overall structure will be the same for all specialized
implementations given they ultimately must all perform the same process,
the generation code itself is abstracted to simplify new additions.
Other than typical function declaration and related setup, all that will
be required for new specialized implementations is to populate the state
of the blocksCompressor struct with the relevant registers and functions
from a new and call its generate method.

The generation code is implemented via a separate internal module to
ensure the specific version of avo used to generate the code is pinned
and therefore entirely reproducible without adding an otherwise
unnecessary dependency to the main public module.

The assembly code can be generating by running `go generate` in the _asm
directory.  It can then be compared against the generated code included
in the commit.