netsync: Name received msg handlers with On, not Queue
This matches the naming used by the public peer package and (because it
implements the peer interface) the server peer.

More importantly, the "Queue" naming is already used by the server/peer
package to push outgoing messages to peers (e.g. QueueMessage and
QueueInventory), and it seemed unnecessarily complex to need to distinguish
the different uses cases of the same names when working on server and netsync
simultaneously.