contrib: Add script to bump assumevalid block.
This adds a script to `contrib/devtools` that queries the main and test
networks with `dcrctl` to determine suitable updated blocks to use for
their assumevalid block and then creates a `git` commit on a new branch
that updates the `chaincfg/mainnetparams.go` and
`chaincfg/testnetparams.go` files accordingly.  The commit description
includes the relevant details and instructions for others to verify the
results.