blockchain: Fix return amts test for Go 1.23+.
This updates the test for ticket return calculation amounts to work
properly with Go 1.23+.

Specifically, it updates the default ticket commitment script that is
used when creating the mock minimal ouputs that the tests then modify to
be an actual commitment script.  Note that this is solely a test issue.

It appears that an invalid mock ticket commitment script was specified
when the test was created, but since the tests only deal with the return
amounts which are explicitly modified in the mock script to the values
the test data requires, it did not previously matter.

However, the code that updates the mock script for the tests
inadvertently relied on undocumented behavior in hex.DecodeString
returning a slice with a larger capacity which allowed the slice to be
extended when setting the appropriate amounts.  However, that is no
longer the case as of Go 1.23+ which uncovered the otherwise invalid
script.