certgen: Prepare v1.2.0.
This updates the certgen module dependencies and serves as a base for
certgen/v1.2.0.

The updated direct dependencies in this commit are as follows:

- golang.org/x/net@v0.28.0

The updated indirect dependencies in this commit are as follows:

- golang.org/x/crypto@v0.26.0
- golang.org/x/sys@v0.23.0
- golang.org/x/text@v0.17.0

The full list of updated dependencies since the previous certgen/v1.1.3
release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.