secp256k1: Tighten max magnitudes in comments.
The FieldVal type was originally an internal only type and the comments
regarding the magnitudes were written from that perspective along with
the the understanding that the stated maximum magnitudes referred to the
max results during normalization.

However, now that the type is external, it was pointed out that the
magnitude comments are misleading since it is not explicitly made clear
anywhere that they refer to what ultimately happens during normalization
as opposed to the specific methods in question.

In order to prevent potential confusion, this updates all of the
comments regarding max magnitudes to be framed in terms of an external
perspective and the specific method they are commenting on such that it
ultimately ensures normalization is guaranteed not to exceed that actual
internal maximum magnitude.