[release-v2.0] mixclient: Sort roots for slot assignment
The purpose of the SR (slot reservation) mix is to determine unique anonymous
slot positions for the DC-net (xor) mix.  However, there was a missing check
by non-root-solving peers that the published roots are in the proper order.
This could result in a malicious actor publishing roots in a different order,
disrupting mixes by causing non-root-solving peers to unknowingly perform the
DC-net in the wrong slots, and causing the wrong peers to be blamed for
disrupting the mix.

Prevent this by requiring roots to be sorted before slots are discovered.
This also matches the root-sorting that occurred in the older client-server
design.

This change is not compatible with previous mixclient versions, as the DC-net
mix would occur with peers writing messages into different slots.  The pairing
version included in PR messages is updated to reflect this.