mixing: Prepare v0.4.2.
This updates the mixing module dependencies and serves as a base for
mixing/v0.4.2.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/crypto/blake256@v1.1.0

The full list of updated direct dependencies since the previous
mixing/v0.4.1 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.