[release-v2.0] main: Use backported mixing updates.
This updates the 2.0 release branch to use the latest version of the
mixing module which includes various fixes to make decentralized mixing
more robust against misbehaving peers.  All participants will need to
update to the latest version to achieve the maximum anonymity set.

In particular, the following updated module version is used:

- github.com/decred/dcrd/mixing@v0.4.2