mixing: Prepare v0.5.0.
This updates the mixing module dependencies and serves as a base
for mixing/v0.5.0.

The updated direct dependencies in this commit are as follows:

- decred.org/cspp/v2@2.4.0
- github.com/decred/dcrd/crypto/rand@v1.0.1
- golang.org/x/crypto@v0.33.0
- golang.org/x/sync@v0.11.0

The updated indirect dependencies in this commit are as follows:

- golang.org/x/sys@v0.30.0

The full list of updated dependencies since the previous mixing/v0.4.2
release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.