netsync: Always request initial non-mining state.
Currently, the existing logic to avoid requesting the initial mining
state is slightly incorrect since the initial state was extended to
include additional data that is not specifically related to the mining
state.  Specifically, it currently does not request any initial state at
all when the when the no mining state synchronization flag is set
instead of only avoiding the data specific to the mining state.

Since the flag is not set by default, most nodes would not encounter
this condition, but the result for any nodes with the no mining sync
flag set is that any unmined treasury spends that are currently being
voted on would not immediately be accessible to nodes initially coming
online.

This resovles the aforementioned by modifying the relevant logic to only
avoid requesting specifically the mining state.

While here, it also changes the logic to prevent duplicate initial state
requests to use a sync primitive that will ultimately allow requests to
be run in parallel instead of requring them all to go through a single
goroutine to protect concurrent access.