mixclient: Error out of run if gen/ke errors
If message (address) generation errors, or any other unexpected error occurs
during the creation of the KE/RS messages, we can't reasonably continue.
Previously this situation would only be reported with a debug log, which could
easily go unnoticed.  Abort the run and log at error level.

If the error occurred during gen, it is possible that the dcMsg was not
properly allocated.  This could result in a out-of-bounds panic later during
the XOR DC-net stage while creating the mix vectors.