main: Respect GOMEMLIMIT environment variable.
This modifies the logic that sets the soft upper memory limit to respect
the GOMEMLIMIT environment variable when it is set so long as it is set
to a value that is higher than the calculated minimum requirement.  This
allows system admins to increase the limit if they are comfortable
trading off more memory for performance in their environment.  It also
provides a way to resolve any potential issues related to the limit in
the future should it be necessary in between releases.