rpcserver: Update getblockchaininfo sync help.
This updates the help text for the initialblockdownload field of the
getblockchaininfo RPC to more accurately refer to the initial chain sync
mode that it represents since what was previously the initial block
download no longer applies and hasn't for a very long time.

Historically, initial block download consisted of downloading all blocks
one-by-one in a linear fashion and only storing them once they had been
fully verified, so the terminology of initial block download accurately
described the process.

However, the logic related to initial chain syncing was changed long ago
to a much more efficient and robust approach based on downloading the
headers first and then using those to guide the sync process.

The RPC field name was not updated to avoid breaking the API, but its
meaning changed and this updates the text accordingly.