blockchain: Consolidate not in main chain err fmt.
This consolidates the error formatting logic for the various cases when
a block that is not in the main chain is requested either by height or
hash.

Rather than having a single generic error string that represents the
various cases and repeatedly formatting the errors at each call site,
this introduces two individual errors of the appropriate types and moves
the error formatting onto the types.

It also updates the tests to remove isNotInMainChainErr which is a
holdover that is no longer needed due to previous changes, as well as to
ensure the new formatted error strings are the expected output.