netsync: Implement separate mutex for peers.
This is the first in a series of commits that will ultimately convert
all of the code related to the sync manager handling events from remote
peers to synchronous code that makes use of separate mutexes and atomics
to protect concurrent access.

The motivation for the change in architecture follows.

Currently, all operations performed by the sync manager are implemented
via a single event handler goroutine and a single message channel to
protect concurrent access.

The existing implementation has worked well for many years since syncing
currently only happens via a single sync peer.

However, with the goal of moving to a syncing paradigm where there is no
sync peer and data is intelligently downloaded from multiple peers in
parallel, the current implementation is not ideal in a few ways:

- It is difficult to improve parallel processing since everything is
  forced into a single goroutine
- It requires additional plumbing to provide access to any related
  information
- The use of channels significantly hinders dynamically modifying
  requests based on state which is necessary for efficiently syncing
  from multiple peers

Converting all the relevant code to synchronous code addresses the
aforementioned concerns and has some additional benefits.

For example, the new model:

- Ultimately allows more code to run in parallel in the individual peer
  goroutines
- Requires less plumbing for handling events
- Makes the state available to calling code so it can make better
  decisions
- Blocks further reads until the messages are processed by default
- Provides better overall throughput

One notable aspect of the overall change to the architecture is that the
peer read goroutines themselves will now block while processing the
messages by default as opposed to the current model where the messages
are queued up to be handled asynchronously which then allows the peers
to immediately read more data.

This existing async queuing approach requires care to prevent malicious
behavior as is evident by several of the handlers (for example, blocks,
transactions, and mix messages) currently manually implementing blocking
behavior by using channels that are notified once the events are
eventually processed by the async handler.

Moreover, since all of the events from all peers all currently go
through the single sync manager goroutine, the aforementioned messages
often end up blocking not only until the message itself is processed,
but also until all of the other messages from all of the other peers
that arrived first are processed.

The existing model also results in further induced delays because most
of the messages result in requests to the peer in response which
obviously can't be sent until the message is processed.

So, while at first glance the lack of read pipelining by default might
raise some potential questions about extra latency, in practice, the new
model will typically result in faster overall processing despite that.
Further, it permits much more efficient per-message targeted pipelining
if any major bottlenecks were to arise in new messages.

With all of that in mind, this commit starts the conversion by
introducing a separate mutex to protect the peers map and updates the
code to protect all references to the map with that mutex.

This will allow future commits to methodically refactor the various
operations without introducing races.