netsync: Separate initial hdr sync done handling.
Currently, the logic that transitions from the initial header sync to
the initial chain sync is quite buried in a conditional in the header
handling routine which makes it somewhat challenging to find and follow
for those not familiar with the code.  It also makes it more difficult
to update, particularly if more than a single sync peer for the initial
header sync is implemented, since it isn't concurrent safe and other
code around it relies on local state it updates.

This aims to improve the aforementioned by separating the handling for
when the manager determines the initial header sync process completes
into a separate method named onInitialHeaderSyncDone.

The main benefits are:

- It is much easier to find and follow the overall sequence of
  events that take place during the initial sync process
- It is more consistent the existing onInitialChainSyncDone method that
  is invoked when the initial chain sync process completes
- It will allow future commits to more easily add additional handling
  and make it concurrent safe

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.