netsync: Track if a peer serves data.
A peer that serves data is not necessarily a sync peer candidate, but
serving data is a necessary condition for being a sync peer candidate.

This helps make that distinction clear by adding a new flag to the Peer
struct that specifies whether or not the peer serves data and updates a
few places in the code that only rely on serving data to use the new
flag instead of overloading the meaning of the sync candidate flag.

It also has the benefit of being concurrent safe, unlike the sync
candidate flag, since whether or not a peer serves data can't change
while sync candidacy can.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.