netsync: Make sync peer candidacy concurrent safe.
This simplifies sync peer candidacy determination and makes it
concurrent safe by removing the sync candidate flag from the Peer struct
and instead basing candidacy on the new immutable serves data flag and
imposing the other necessary conditions when choosing the best sync
peer.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.