netsync: Use atomic for sync height.
This modifies the sync height to use an atomic.Int64 instead of a mutex
to protect concurrent access and introduces a new method that atomically
and conditionally sets it to a higher value.

This approach makes it both harder to misuse and quite a bit faster due
to using non-blocking optimistic locking.

It also adds a test to ensure the new method works as expected including
under concurrent access.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.