netsync: Use atomic for is current flag.
This modifies the isCurrent flag to use an atomic.Bool instead of a
mutex to protect concurrent access.

This approach makes it both harder to misuse and quite a bit faster due
to using non-blocking optimistic locking.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.