netsync: Use atomic for headers synced flag.
This changes the headersSynced flag to an atomic.Bool so it is safe to
access concurrently.

It also introduces a new method on the header sync state named
InitialHeaderSyncDone to simplify and improve the readability of the
various places that check its state.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.