netsync: Make sync peer concurrent safe.
This makes access to the sync peer independently concurrent safe by
protecting it with a mutex and updating all of the call sites
appropriately.

It also ensures the methods related to starting the various chain sync
steps from the sync peer are protected by the mutex since their overall
logic relies on its value to determine when they need to be invoked.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.