netsync: Make request maps concurrent safe.
This makes all accesses to the request maps independently concurrent
safe by protecting them with a mutex and updating all of the call sites
accordingly.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.