netsync: Make next block fetch concurrent safe.
This makes all logic related to requesting the next needed blocks from
the sync peer independently concurrent safe by protecting it with a
mutex and updating all of the call sites accordingly.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.