netsync: Make stall timeout concurrent safe.
This modifies the stall timeout logic to make it concurrent safe by
slightly loosening the ordering requirements as it relates to stopping
and resetting the stall timeout.

The current implementation was careful to prevent stale time values from
being delivered on the timer's channel, however, there really is no
reason to worry about that possibility because:

1) The stall timeout logic only depends on the event as opposed to the
   time values
2) The timeout is sufficiently long that it is almost never hit in
   practice
3) It is even more rare that the timer would just happen to be in the
   process of firing (sending to the channel) when data simultaneously
   arrives and resets the timer just before it actually fires
4) Even if all of that were to happen, the only effect is the standard
   stall timeout handling logic which means there are no adverse effects

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.