netsync: Make add peer dc logic synchronous.
This refactors the logic related to handling valid peers disconnecting
from the sync manager out of the event handler since it is now
independently concurrent safe.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.