netsync: Make best ann tracking concurrent safe.
This makes all logic related to tracking the best announced block by
each peer independently concurrent safe by protecting the related fields
with a mutex and updating all of the call sites accordingly.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.