netsync: Make sync transition concurrent safe.
The current approach to detecting when the initial chain sync process is
done is based on determining if the chain wasn't current prior to
processing the block and becomes current after.  That logic is sound so
long as the entire sequence is mutually exclusive, as is currently the
case, but it would result in logic races when processing multiple blocks
concurrently.

Consequently, this reworks the logic related to transitioning to the
initial chain sync done state slightly to be independently concurrent
safe by introducing a new flag and mutex to track its state and then
transitioning based on the flag and status of whether the chain believes
it is current.

It also squashes a very minor long-standing logging bug where the
transition out of the initial chain sync mode would sometimes not log
the last batch of processed blocks for either an extended period or not
at all.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.