netsync: Make transaction handling synchronous.
This refactors the logic related to handling transaction messages from
peers out of the event handler since it is now independently concurrent
safe.

It also slightly modifies the behavior to return the slice of accepted
transactions to the caller so it can announce them as opposed to relying
on a callback to do the announcements.

The announcement callback is removed from the peer notifier interface
since it is no longer needed.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.