netsync: Make mix message handling synchronous.
This refactors the logic related to handling mix messages from peers out
of the event handler since it is now independently concurrent safe.

It also slightly modifies the behavior to return the slice of accepted
messages to the caller so it can announce them as opposed to relying on
a callback to do the announcements.

The entire peer notifier interface and containing file is removed since
it is no longer needed.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.