netsync: Rename event handler to stall handler.
This renames the event handler goroutine to stallHandler to more
accurately reflect its only remaining purpose after all of the recent
changes to move everything else out of the event handler.

This is a part of the overall effort to convert the code related to
handling the various sync manager events to synchronous code that runs
in the various caller goroutines.