netsync: Split initial header and chain sync.
This splits the logic for starting the initial header sync and initial
chain sync processes into separate methods and also improves the warning
when there are no sync peers to be concurrent safe as well as to prevent
multiple unnecessary warnings.

The primary motivation for this change is to make it possible to
independently make the state transitions between the modes concurrent
safe in upcoming future commits.

Looking even further ahead, it partially paves the way to eventually
removing the entire notion of a chain sync peer (but not header sync
peer) entirely.

However, it also improves readability and has the nice minor benefit of
slightly improved efficiency when peers connect and disconnect since it
allows the logic that does not apply during the initial headers sync to
more easily be skipped without needing to check the flag multiple times.